port to chrome
error when packaging crx with chrome, have to load unpacked. tested on edge chromium and also works. For publishing in chrome web store a $5 fee needs to be paid. However company is set to whitelist extensions therefore even if I paid the extension could not be downloaded. For edge you could manually load it unpacked but main target audience was chrome. Lost of time but great coding practice.